Combined traits for group chats and group chat messages
