+++ akka-bbb-apps/src/main/scala/org/bigbluebutton/core/running/MeetingActor.scala
@@ -362,11 +362,8 @@ class MeetingActor(
     // sync all presentations
     presentationPodsApp.handleSyncGetPresentationPods(state, liveMeeting, msgBus)
 
-    // sync all group chats
-    groupChatApp.handleSyncGetGroupChats(state, liveMeeting, msgBus)
-
-    // sync all group chat messages
-    groupChatApp.handleSyncGetGroupChatMsgs(state, liveMeeting, msgBus)
+    // sync all group chats and group chat messages
+    groupChatApp.handleSyncGetGroupChatsInfo(state, liveMeeting, msgBus)
 
     // TODO send all lock settings
     // TODO send all screen sharing info
